server/pyproject.toml
[tool.poetry]
name = "shift"
version = "0.1.0"
description = "My new django project"
authors = ["ZZaRRyaDD <example@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.5"
Pillow = "9.5.0"
django = "4.2.1"
redis = "4.5.5"
django-redis = "5.2.0"
celery = {extras = ["redis"], version = "5.2.7"}
flower = "1.2.0"
django-model-utils = "4.3.1"
djangorestframework = "3.14.0"
django-cors-headers = "3.14.0"
psycopg2-binary = "2.9.6"
gunicorn = "20.1.0"
django-extensions = "3.2.1"
dj-database-url = "1.3.0"
djangorestframework-simplejwt = "4.8.0"
django-filter = "23.2"

[tool.poetry.group.dev.dependencies]
django-stubs = "1.16.0"
pytest = "7.3.1"
pytest-sugar = "0.9.7"
pytest-django = "4.5.2"
djangorestframework-stubs = "1.10.0"
flake8 = "4.0.1"
flake8-isort = "4.2.0"
factory-boy = "3.2.1"
django-debug-toolbar = "3.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
